+++ src/constants/transactions.constants.ts
@@ -1,4 +1,5 @@
 export const CANCEL_TRANSACTION_TOAST_ID = 'cancel-transaction-toast';
+export const ERROR_SIGNING_TOAST_ID = 'error-signing-toast';
 export const AVERAGE_TX_DURATION_MS = 6000;
 export const CROSS_SHARD_ROUNDS = 5;
 export const TRANSACTIONS_STATUS_POLLING_INTERVAL_MS = 90 * 1000; // 90sec