+++ src/store/slices/toast/toastSlice.ts
@@ -1,18 +1,17 @@
 import { StateCreator } from 'zustand/vanilla';
 import { StoreType, MutatorsIn } from 'store/store.types';
-import { IToastsSliceState } from './toastSlice.types';
+import { ToastsSliceType } from './toastSlice.types';
 
-export const initialState: IToastsSliceState = {
+export const initialState: ToastsSliceType = {
   customToasts: [],
-  transactionToasts: [],
-  toastProgress: {}
+  transactionToasts: []
 };
 
 function getToastSlice(): StateCreator<
   StoreType,
   MutatorsIn,
   [],
-  IToastsSliceState
+  ToastsSliceType
 > {
   return () => initialState;
 }