+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- [Added `FormatAmountController` and `TransactionsTableController` tests](https://github.com/multiversx/mx-sdk-dapp-core/pull/83)
+- [Cleanup transactions and tracked transactions logic](https://github.com/multiversx/mx-sdk-dapp-core/pull/82)
+- [Enhance toast progress](https://github.com/multiversx/mx-sdk-dapp-core/pull/81)
+- [Fixed react type mismatch](https://github.com/multiversx/mx-sdk-dapp-core/pull/80)
+- [Removed React hooks from `TransactionsTableController` `processTransaction` method](https://github.com/multiversx/mx-sdk-dapp-core/pull/79)
+- [Added transaction value in controller processing](https://github.com/multiversx/mx-sdk-dapp-core/pull/78)
+- [Enhance handle sign error](https://github.com/multiversx/mx-sdk-dapp-core/pull/77)
+- [Added transaction accounts in controller processing](https://github.com/multiversx/mx-sdk-dapp-core/pull/76)
 - [Fixed track transactions web socket re-trigger](https://github.com/multiversx/mx-sdk-dapp-core/pull/75)
 - [Added devtools action names](https://github.com/multiversx/mx-sdk-dapp-core/pull/74)
 - [Added transaction display info support](https://github.com/multiversx/mx-sdk-dapp-core/pull/73)